+++ sdk/core/azure-core-amqp/src/main/java/com/azure/core/amqp/implementation/CreateSessionOptions.java
@@ -1,0 +1,30 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.core.amqp.implementation;
+
+import com.azure.core.annotation.Immutable;
+
+/**
+ * An option bag to provide configuration required to create an AMQP session.
+ */
+@Immutable
+public class CreateSessionOptions {
+    private final boolean distributedTransactionsSupport;
+
+    /**
+     * Constructor to create {@link CreateSessionOptions}.
+     * @param distributedTransactionsSupport if AMQP session supports distributed transaction across different entities.
+     */
+    public CreateSessionOptions(boolean distributedTransactionsSupport) {
+        this.distributedTransactionsSupport = distributedTransactionsSupport;
+    }
+
+    /**
+     * Determine is distributed transactions are supported across different entities.
+     * @return true if distributed transactions across different entities are supported.
+     */
+    public boolean isDistributedTransactionsSupported() {
+        return this.distributedTransactionsSupport;
+    }
+}