+++ sdk/core/azure-core-amqp/src/main/java/com/azure/core/amqp/implementation/ReactorConnection.java
@@ -179,7 +179,7 @@ public class ReactorConnection implements AmqpConnection {
      * {@inheritDoc}
      */
     @Override
-    public Mono<AmqpSession> createSession(String sessionName, boolean coordinatorRequired) {
+    public Mono<AmqpSession> createSession(String sessionName, boolean distributedTransactionsSupport) {
         if (isDisposed()) {
             return Mono.error(logger.logExceptionAsError(new IllegalStateException(String.format(
                 "connectionId[%s]: Connection is disposed. Cannot create session '%s'.", connectionId, sessionName))));
@@ -197,7 +197,7 @@ public class ReactorConnection implements AmqpConnection {
                 final Session session = connection.session();
 
                 BaseHandler.setHandler(session, handler);
-                final AmqpSession amqpSession = createSession(key, session, handler, coordinatorRequired);
+                final AmqpSession amqpSession = createSession(key, session, handler, distributedTransactionsSupport);
                 final Disposable subscription = amqpSession.getEndpointStates()
                     .subscribe(state -> {
                     }, error -> {
@@ -236,15 +236,15 @@ public class ReactorConnection implements AmqpConnection {
      * @param sessionName Name of the AMQP session.
      * @param session The reactor session associated with this session.
      * @param handler Session handler for the reactor session.
-     * @param coordinatorRequired If coordinator is required for this session.
+     * @param distributedTransactionsSupport If distributed transaction across entities is required for this session.
      *
      * @return A new instance of AMQP session.
      */
     protected AmqpSession createSession(String sessionName, Session session, SessionHandler handler,
-        boolean coordinatorRequired) {
+        boolean distributedTransactionsSupport) {
         return new ReactorSession(session, handler, sessionName, reactorProvider, handlerProvider,
             getClaimsBasedSecurityNode(), tokenManagerProvider, messageSerializer, connectionOptions.getRetry(),
-            coordinatorRequired);
+            new CreateSessionOptions(distributedTransactionsSupport));
     }
 
     /**