+++ sdk/core/azure-core-amqp/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release History
 
 ## 2.1.0-beta.1 (Unreleased)
+### New Features
+- Added support for distributed transactions across entities. 
 
 ## 2.0.3 (2021-03-09)